refactor: "there is a DimensionStructure domain object" step is refactored to a common project
