+++ src/MasterData/MasterData.BusinessLogic.Features.Tests/MasterData.BusinessLogic.Features.Tests.csproj
@@ -52,6 +52,7 @@
         <ProjectReference Include="..\..\Utils\MasterDataTestHelper\MasterDataTestHelper.csproj"/>
         <ProjectReference Include="..\MasterData.BusinessLogic.Implementations\MasterData.BusinessLogic.Implementations.csproj"/>
         <ProjectReference Include="..\MasterData.BusinessLogic.Interfaces\MasterData.BusinessLogic.Interfaces.csproj"/>
+        <ProjectReference Include="..\MasterData.StepDefinitions\MasterData.StepDefinitions.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
@@ -116,6 +117,12 @@
 
     <ItemGroup>
         <Folder Include="StepDefinitions\Utils"/>
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Update="specflow.json">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </None>
     </ItemGroup>
 
 </Project>
