+++ src/MasterData/MasterData.StepDefinitions.old/MasterData.StepDefinitions.old.csproj
@@ -1,0 +1,23 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+        <AssemblyName>DigitalLibrary.MasterData.StepDefinitions</AssemblyName>
+        <RootNamespace>DigitalLibrary.MasterData.StepDefinitions</RootNamespace>
+        <LangVersion>default</LangVersion>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Bogus" Version="33.0.2" />
+      <PackageReference Include="SpecFlow" Version="3.7.38" />
+      <PackageReference Include="SpecFlow.Tools.MsBuild.Generation" Version="3.7.38" />
+      <PackageReference Include="SpecFlow.xUnit" Version="3.7.38" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\Utils\Guards\Guards.csproj" />
+      <ProjectReference Include="..\MasterData.DomainModel.Interfaces\MasterData.DomainModel.Interfaces.csproj" />
+      <ProjectReference Include="..\MasterData.DomainModel\MasterData.DomainModel.csproj" />
+    </ItemGroup>
+
+</Project>
