+++ src/MasterData/MasterData.Web.Api.Features.Tests/DimensionStructure/AddAsync.feature
@@ -9,9 +9,6 @@ Feature: DimensionStructure REST endpoint - AddAsync method
     Given there is a DimensionStructure domain object
       | Field    | Value |
       | Key      | ds    |
-      | Name     | asd   |
-      | Desc     | asd   |
-      | ISActive | 1     |
     And DimensionStructure is sent to the Add endpoint
       | Field     | Value     |
       | Key       | ds        |
